tg_digest/requirements.txt
telethon>=1.43.2
groq>=1.2.0
python-telegram-bot>=22.7
python-dotenv>=1.0.0
schedule>=1.2.0
